+++ src/test/java/hello/proxy/pureProxy/decorator/code/Component.java
@@ -1,4 +1,4 @@
-package hello.proxy.decorator.code;
+package hello.proxy.pureProxy.decorator.code;
 
 public interface Component {
